# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chat-App",
    "description": "Chat-App Server Project",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/localhost:3000",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User create service",
        "description": "",
        "operationId": "addUser",
        "parameters": [
          {
            "in": "body",
            "name": "register",
            "description": "The user to create.",
            "schema": {
              "type": "object",
              "required": [
                "full_name",
                "email",
                "password"
              ],
              "properties": {
                "full_name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/messages/send-message": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Send message service",
        "description": "",
        "operationId": "sendMessage",
        "parameters": [
          {
            "in": "body",
            "name": "Send message",
            "description": "",
            "schema": {
              "type": "object",
              "required": [
                "roomId",
                "senderId",
                "receiverId",
                "message"
              ],
              "properties": {
                "roomId": {
                  "type": "string"
                },
                "senderId": {
                  "type": "string"
                },
                "receiverId": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/messages/list-message-by-room-id": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "List message service",
        "description": "",
        "operationId": "listMessage",
        "parameters": [
          {
            "in": "body",
            "name": "Send message",
            "description": "",
            "schema": {
              "type": "object",
              "required": [
                "room_id"
              ],
              "properties": {
                "room_id": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  }
}